[Performance Improvement] Changed from rest API to GraphQL for repository listing
The repository listing now is really slow, that's partly because the
navikt has a lot of repos, and becuase the rest API is pulling in alot
more information than necessary.

To fix the latter, I've implemented a graphql repository list function
that's significantly faster, it's also able to paginate with 100 repos
at a time which reduces the amount of requests.

GraphQL fixes the existing issue, because I can request exactly the
fields needed in a faster and more compact way. Which takes the wait
from 20 seconds down to 11, in my testing.