[Feature] Rewrote the frontend in Vue
For now it's just a basic rewrite. It started as just a search function,
but then I needed to get the repositories using the existing
functions... So I just rewrote the existing request functions aswell,
and then I finished the rest of the functionality and stripped out the
rest of the JS.

Now it's only vue and in my opinion the code reads better as lots of the
underlying logic the previous JS did is handeled by Vue or atleast
written in a more clear way.

As for features, it's on par with the previous release with these
changes:
- Search, you can now search using the name, organization or the URL
- Errors, errors are now setup as a list to accomodate the various
problems that could be going on
- Progress animation with status
- Form validation and gradual apperance of the required elements of the
web page