[Performance Improvement] Improved repository listing speed
So the issue is, navikt has over 900 repos and growing, even though my
graphQL solution made it faster, the fetching is still really slow. That
irritated me to the point, where I tried a lot of things to make it
better, but I ended on this solution.

Here i'm just using graphQL's ability to sort the repos, and getting
them in a descending order by Creation date. With that fact i can get
all the repos once, and just add to the list of repos as repos created
at a newer date appear.

Now! If a repo is deleted or there is a mismatch in some way all of the
repos are fetched again, just like when it starts up for the first time.

This makes the overal experience a lot more snappy while also reducing
the amount of quries we do towards github.

My solution is based on a struct containing all of the known
repositories, aswell as the date of the repository created most
reacently. The struct also has a R/W mutex so concurrent access won't mess
it up.