grpc/dispatchserver: remove online cluster if `invalidateHistoric` fails
If the `invalidateHistoric` method returns an error (e.g. if a database
connection is unavailable), the GRPC stream will end and disconnect the
client.

The previous implementation did not remove the cluster from
the `onlineClustersMap`, resulting in a deadlocked loop where the
client would attempt to reconnect only to be rejected due to already
being marked as online.

We fix this by also always removing the cluster from the `onlineClustersMap`
if the `invalidateHistoric` method returns an error.